+++ src/components/sequence/items/index.js
@@ -43,6 +43,7 @@ import SmartExposureItem from './SmartExposureItem.vue';
 import RunAutofocusItem from './RunAutofocusItem.vue';
 import SequentialContainerItem from './SequentialContainerItem.vue';
 import CenterItem from './CenterItem.vue';
+import SlewScopeToRaDecItem from './SlewScopeToRaDecItem.vue';
 import DeepSkyObjectContainerItem from './DeepSkyObjectContainerItem.vue';
 import WarmCameraItem from './WarmCameraItem.vue';
 import MoonAltitudeConditionItem from './MoonAltitudeConditionItem.vue';
@@ -139,7 +140,7 @@ export const ITEM_COMPONENTS = {
   'NINA.Sequencer.SequenceItem.Switch.SetSwitchValue': SetSwitchValueItem,
   'NINA.Sequencer.SequenceItem.Telescope.SetTracking': SetTrackingItem,
   'NINA.Sequencer.SequenceItem.Telescope.SlewScopeToAltAz': SlewScopeToAltAzItem,
-  'NINA.Sequencer.SequenceItem.Telescope.SlewScopeToRaDec': CenterItem,
+  'NINA.Sequencer.SequenceItem.Telescope.SlewScopeToRaDec': SlewScopeToRaDecItem,
   'NINA.Sequencer.SequenceItem.Platesolving.CenterAndRotate': CenterAndRotateItem,
   'NINA.Sequencer.SequenceItem.Platesolving.SolveAndSync': RunAutofocusItem,
   'NINA.Sequencer.SequenceItem.Expressions.GlobalConstant': ExpressionVariableItem,